third_party/WebKit/LayoutTests/fast/dom/cross-frame-accessor-throw-expected.txt

This is a testharness.js-based test.
PASS Check that DOM exception is created in called function's context. 
FAIL Check that DOM exception is created in setter's context. assert_equals: expected [stringifying object threw TypeError: Illegal invocation with type object] but got [stringifying object threw TypeError: Illegal invocation with type object]
PASS Check that native exception is created in setter's context. 
Harness: the test ran to completion.
